refactor: unify id params to int and thread context.Context

+++ core-service/internal/models/cart.go
@@ -4,7 +4,7 @@ import "time"
 
 type Cart struct {
 	UserID    int       `gorm:"column:user_id;primaryKey"`
-	ProductID uint64    `gorm:"column:product_id;primaryKey"`
+	ProductID int       `gorm:"column:product_id;primaryKey"`
 	Product   Product   `gorm:"foreignKey:ProductID;references:ID"`
 	Qty       int       `gorm:"column:qty"`
 	CreatedAt time.Time `gorm:"column:created_at"`
